move ct_export2cairo content in ct_export2pdf and remove it
